refactoring company details form
